# api/docs/swagger.yaml
definitions:
...
  model.ErrValidation:
    properties:
      invalid_params:
        items:
          $ref: '#/definitions/model.InvalidParam'
        type: array
    type: object
  model.InvalidParam:
    properties:
      name:
        type: string
      reason:
        type: string
    type: object
...
  model.Product:
    properties:
      buy_price:
        type: number
      msrp:
        type: number
      product_code:
        type: string
      product_description:
        type: string
      product_line:
        type: string
      product_name:
        type: string
      product_scale:
        type: string
      product_vendor:
        type: string
      quantity_in_stock:
        type: integer
    type: object
...
info:
  contact:
    email: Martsoftfilmz@gmail.com
    name: Martsoft Inc
  description: This is the API E-Commerce businesses.
  title: Martsoft Inc E-Commerce API 2.0
  version: "1.0"
paths:
...
  /products/delete:
    delete:
      description: Deletes a single Product
      parameters:
      - description: product_code mandatory
        in: query
        name: product_code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrValidation'
      summary: Deletes a single Product
      tags:
      - Products
  /products/get:
    get:
      description: Fetch a single Product
      parameters:
      - description: product_code mandatory
        in: query
        name: product_code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrValidation'
      summary: Retrieve a single Product
      tags:
      - Products
  /products/products:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrValidation'
      summary: Retrieve all Payments
      tags:
      - Products
  /products/update:
    patch:
      description: updates a single Product
      parameters:
      - description: quantity_in_stock
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/model.Product'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrValidation'
      summary: Updates a  Product
      tags:
      - Products
schemes:
- http
swagger: "2.0"